+++ pkg/application/doguInstallationUseCase_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/cloudogu/k8s-blueprint-operator/pkg/domain/ecosystem"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"k8s.io/apimachinery/pkg/api/resource"
 	"testing"
 	"time"
 )
@@ -22,6 +23,7 @@ var postgresqlQualifiedName = common.QualifiedDoguName{
 	SimpleName: "postgresql",
 }
 
+// TODO Add test for proxy and volume actions
 func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 	t.Run("action none", func(t *testing.T) {
 		// given
@@ -40,7 +42,7 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 				Version:           version3211,
 				InstallationState: domain.TargetStatePresent,
 			},
-			NeededAction: domain.ActionNone,
+			NeededActions: []domain.Action{domain.ActionNone},
 		}, &ecosystem.DoguInstallation{
 			Name:    postgresqlQualifiedName,
 			Version: version3211,
@@ -51,9 +53,16 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 	})
 
 	t.Run("action install", func(t *testing.T) {
+		volumeSize := resource.MustParse("2Gi")
+		bodySize := resource.MustParse("2G")
+		config := ecosystem.ReverseProxyConfig{
+			MaxBodySize:      &bodySize,
+			RewriteTarget:    "/",
+			AdditionalConfig: "additional",
+		}
 		doguRepoMock := newMockDoguInstallationRepository(t)
 		doguRepoMock.EXPECT().
-			Create(testCtx, ecosystem.InstallDogu(postgresqlQualifiedName, version3211)).
+			Create(testCtx, ecosystem.InstallDogu(postgresqlQualifiedName, version3211, &volumeSize, config)).
 			Return(nil)
 
 		sut := NewDoguInstallationUseCase(nil, doguRepoMock, nil)
@@ -69,11 +78,13 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 					InstallationState: domain.TargetStateAbsent,
 				},
 				Expected: domain.DoguDiffState{
-					Namespace:         "official",
-					Version:           version3211,
-					InstallationState: domain.TargetStatePresent,
+					Namespace:          "official",
+					Version:            version3211,
+					InstallationState:  domain.TargetStatePresent,
+					MinVolumeSize:      &volumeSize,
+					ReverseProxyConfig: config,
 				},
-				NeededAction: domain.ActionInstall,
+				NeededActions: []domain.Action{domain.ActionInstall},
 			},
 			nil,
 			domain.BlueprintConfiguration{},
@@ -95,8 +106,8 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 		err := sut.applyDoguState(
 			testCtx,
 			domain.DoguDiff{
-				DoguName:     "postgresql",
-				NeededAction: domain.ActionUninstall,
+				DoguName:      "postgresql",
+				NeededActions: []domain.Action{domain.ActionUninstall},
 			},
 			&ecosystem.DoguInstallation{
 				Name:    postgresqlQualifiedName,
@@ -129,7 +140,7 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 				Expected: domain.DoguDiffState{
 					Version: version3212,
 				},
-				NeededAction: domain.ActionUpgrade,
+				NeededActions: []domain.Action{domain.ActionUpgrade},
 			},
 			dogu,
 			domain.BlueprintConfiguration{},
@@ -157,7 +168,7 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 				Expected: domain.DoguDiffState{
 					Version: version3211,
 				},
-				NeededAction: domain.ActionDowngrade,
+				NeededActions: []domain.Action{domain.ActionDowngrade},
 			},
 			dogu,
 			domain.BlueprintConfiguration{},
@@ -166,6 +177,234 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 		// then
 		require.ErrorContains(t, err, getNoDowngradesExplanationTextForDogus())
 		assert.Equal(t, version3212, dogu.Version)
+	})
+
+	t.Run("action update volume size", func(t *testing.T) {
+		volumeSize := resource.MustParse("2Gi")
+		expectedVolumeSize := resource.MustParse("3Gi")
+		expectedDogu := &ecosystem.DoguInstallation{
+			Name:          postgresqlQualifiedName,
+			MinVolumeSize: &expectedVolumeSize,
+		}
+
+		dogu := &ecosystem.DoguInstallation{
+			Name:          postgresqlQualifiedName,
+			MinVolumeSize: &volumeSize,
+		}
+
+		doguRepoMock := newMockDoguInstallationRepository(t)
+		doguRepoMock.EXPECT().Update(testCtx, expectedDogu).Return(nil)
+
+		sut := NewDoguInstallationUseCase(nil, doguRepoMock, nil)
+
+		// when
+		err := sut.applyDoguState(
+			testCtx,
+			domain.DoguDiff{
+				DoguName: "postgresql",
+				Expected: domain.DoguDiffState{
+					MinVolumeSize: &expectedVolumeSize,
+				},
+				NeededActions: []domain.Action{domain.ActionUpdateDoguResourceMinVolumeSize},
+			},
+			dogu,
+			domain.BlueprintConfiguration{},
+		)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("action update proxy body size", func(t *testing.T) {
+		proxyBodySize := resource.MustParse("2G")
+		expectedProxyBodySize := resource.MustParse("3G")
+		expectedDogu := &ecosystem.DoguInstallation{
+			Name: postgresqlQualifiedName,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				MaxBodySize: &expectedProxyBodySize,
+			},
+		}
+
+		dogu := &ecosystem.DoguInstallation{
+			Name: postgresqlQualifiedName,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				MaxBodySize: &proxyBodySize,
+			},
+		}
+
+		doguRepoMock := newMockDoguInstallationRepository(t)
+		doguRepoMock.EXPECT().Update(testCtx, expectedDogu).Return(nil)
+
+		sut := NewDoguInstallationUseCase(nil, doguRepoMock, nil)
+
+		// when
+		err := sut.applyDoguState(
+			testCtx,
+			domain.DoguDiff{
+				DoguName: "postgresql",
+				Expected: domain.DoguDiffState{
+					ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+						MaxBodySize: &expectedProxyBodySize,
+					},
+				},
+				NeededActions: []domain.Action{domain.ActionUpdateDoguProxyBodySize},
+			},
+			dogu,
+			domain.BlueprintConfiguration{},
+		)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("action update proxy rewrite target", func(t *testing.T) {
+		target := ecosystem.RewriteTarget("")
+		expectedTarget := ecosystem.RewriteTarget("/")
+		expectedDogu := &ecosystem.DoguInstallation{
+			Name: postgresqlQualifiedName,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				RewriteTarget: expectedTarget,
+			},
+		}
+
+		dogu := &ecosystem.DoguInstallation{
+			Name: postgresqlQualifiedName,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				RewriteTarget: target,
+			},
+		}
+
+		doguRepoMock := newMockDoguInstallationRepository(t)
+		doguRepoMock.EXPECT().Update(testCtx, expectedDogu).Return(nil)
+
+		sut := NewDoguInstallationUseCase(nil, doguRepoMock, nil)
+
+		// when
+		err := sut.applyDoguState(
+			testCtx,
+			domain.DoguDiff{
+				DoguName: "postgresql",
+				Expected: domain.DoguDiffState{
+					ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+						RewriteTarget: expectedTarget,
+					},
+				},
+				NeededActions: []domain.Action{domain.ActionUpdateDoguProxyRewriteTarget},
+			},
+			dogu,
+			domain.BlueprintConfiguration{},
+		)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("action update proxy additional config", func(t *testing.T) {
+		additionalConfig := ecosystem.AdditionalConfig("")
+		expectedAdditionalConfig := ecosystem.AdditionalConfig("snippet")
+		expectedDogu := &ecosystem.DoguInstallation{
+			Name: postgresqlQualifiedName,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				AdditionalConfig: expectedAdditionalConfig,
+			},
+		}
+
+		dogu := &ecosystem.DoguInstallation{
+			Name: postgresqlQualifiedName,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				AdditionalConfig: additionalConfig,
+			},
+		}
+
+		doguRepoMock := newMockDoguInstallationRepository(t)
+		doguRepoMock.EXPECT().Update(testCtx, expectedDogu).Return(nil)
+
+		sut := NewDoguInstallationUseCase(nil, doguRepoMock, nil)
+
+		// when
+		err := sut.applyDoguState(
+			testCtx,
+			domain.DoguDiff{
+				DoguName: "postgresql",
+				Expected: domain.DoguDiffState{
+					ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+						AdditionalConfig: expectedAdditionalConfig,
+					},
+				},
+				NeededActions: []domain.Action{domain.ActionUpdateDoguProxyAdditionalConfig},
+			},
+			dogu,
+			domain.BlueprintConfiguration{},
+		)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("should process multiple update actions", func(t *testing.T) {
+		volumeSize := resource.MustParse("2Gi")
+		expectedVolumeSize := resource.MustParse("3Gi")
+		proxyBodySize := resource.MustParse("2G")
+		expectedProxyBodySize := resource.MustParse("3G")
+		target := ecosystem.RewriteTarget("")
+		expectedTarget := ecosystem.RewriteTarget("/")
+		additionalConfig := ecosystem.AdditionalConfig("")
+		expectedAdditionalConfig := ecosystem.AdditionalConfig("snippet")
+		expectedDogu := &ecosystem.DoguInstallation{
+			Name:          postgresqlQualifiedName,
+			Version:       version3212,
+			MinVolumeSize: &expectedVolumeSize,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				MaxBodySize:      &expectedProxyBodySize,
+				RewriteTarget:    expectedTarget,
+				AdditionalConfig: expectedAdditionalConfig,
+			},
+		}
+
+		dogu := &ecosystem.DoguInstallation{
+			Name:          postgresqlQualifiedName,
+			Version:       version3211,
+			MinVolumeSize: &volumeSize,
+			ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+				MaxBodySize:      &proxyBodySize,
+				RewriteTarget:    target,
+				AdditionalConfig: additionalConfig,
+			},
+		}
+
+		doguRepoMock := newMockDoguInstallationRepository(t)
+		doguRepoMock.EXPECT().Update(testCtx, expectedDogu).Return(nil)
+
+		sut := NewDoguInstallationUseCase(nil, doguRepoMock, nil)
+
+		// when
+		err := sut.applyDoguState(
+			testCtx,
+			domain.DoguDiff{
+				DoguName: "postgresql",
+				Expected: domain.DoguDiffState{
+					Version:       version3212,
+					MinVolumeSize: &expectedVolumeSize,
+					ReverseProxyConfig: ecosystem.ReverseProxyConfig{
+						MaxBodySize:      &expectedProxyBodySize,
+						RewriteTarget:    expectedTarget,
+						AdditionalConfig: expectedAdditionalConfig,
+					},
+				},
+				NeededActions: []domain.Action{
+					domain.ActionUpgrade,
+					domain.ActionUpdateDoguProxyAdditionalConfig,
+					domain.ActionUpdateDoguProxyBodySize,
+					domain.ActionUpdateDoguProxyRewriteTarget,
+					domain.ActionUpdateDoguResourceMinVolumeSize,
+				},
+			},
+			dogu,
+			domain.BlueprintConfiguration{},
+		)
+
+		// then
+		require.NoError(t, err)
 	})
 
 	t.Run("action SwitchNamespace not allowed", func(t *testing.T) {
@@ -184,7 +423,7 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 				Expected: domain.DoguDiffState{
 					Namespace: "premium",
 				},
-				NeededAction: domain.ActionSwitchDoguNamespace,
+				NeededActions: []domain.Action{domain.ActionSwitchDoguNamespace},
 			},
 			dogu,
 			domain.BlueprintConfiguration{
@@ -214,7 +453,7 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 				Expected: domain.DoguDiffState{
 					Namespace: "premium",
 				},
-				NeededAction: domain.ActionSwitchDoguNamespace,
+				NeededActions: []domain.Action{domain.ActionSwitchDoguNamespace},
 			},
 			dogu,
 			domain.BlueprintConfiguration{
@@ -239,7 +478,7 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 				Expected: domain.DoguDiffState{
 					Namespace: "premium",
 				},
-				NeededAction: "unknown",
+				NeededActions: []domain.Action{"unknown"},
 			},
 			nil,
 			domain.BlueprintConfiguration{},
@@ -249,6 +488,28 @@ func TestDoguInstallationUseCase_applyDoguState(t *testing.T) {
 		require.ErrorContains(t, err, "cannot perform unknown action \"unknown\"")
 	})
 
+	t.Run("should fail with no actions", func(t *testing.T) {
+		// given
+		sut := NewDoguInstallationUseCase(nil, nil, nil)
+
+		// when
+		err := sut.applyDoguState(
+			testCtx,
+			domain.DoguDiff{
+				DoguName: "postgresql",
+				Expected: domain.DoguDiffState{
+					Namespace: "premium",
+				},
+				NeededActions: []domain.Action{},
+			},
+			nil,
+			domain.BlueprintConfiguration{},
+		)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "fail because the total amount of actions is zero. actions should contain at least action \"none\" if nothing changed")
+	})
 }
 
 func TestDoguInstallationUseCase_ApplyDoguStates(t *testing.T) {
@@ -293,8 +554,8 @@ func TestDoguInstallationUseCase_ApplyDoguStates(t *testing.T) {
 			StateDiff: domain.StateDiff{
 				DoguDiffs: []domain.DoguDiff{
 					{
-						DoguName:     "postgresql",
-						NeededAction: domain.ActionNone,
+						DoguName:      "postgresql",
+						NeededActions: []domain.Action{domain.ActionNone},
 					},
 				},
 			},
@@ -320,8 +581,8 @@ func TestDoguInstallationUseCase_ApplyDoguStates(t *testing.T) {
 			StateDiff: domain.StateDiff{
 				DoguDiffs: []domain.DoguDiff{
 					{
-						DoguName:     "postgresql",
-						NeededAction: domain.ActionDowngrade,
+						DoguName:      "postgresql",
+						NeededActions: []domain.Action{domain.ActionDowngrade},
 					},
 				},
 			},