+++ pkg/domain/ecosystem/reverseProxyConfig.go
@@ -1,0 +1,9 @@
+package ecosystem
+
+import (
+	"k8s.io/apimachinery/pkg/api/resource"
+)
+
+type BodySize = resource.Quantity
+type RewriteTarget string
+type AdditionalConfig string