+++ pkg/domain/stateDiff.go
@@ -41,11 +41,15 @@ func (diff StateDiff) GetSensitiveDoguConfigDiffsByAction() map[ConfigAction]Sen
 type Action string
 
 const (
-	ActionNone                     = "none"
-	ActionInstall                  = "install"
-	ActionUninstall                = "uninstall"
-	ActionUpgrade                  = "upgrade"
-	ActionDowngrade                = "downgrade"
-	ActionSwitchDoguNamespace      = "dogu namespace switch"
-	ActionSwitchComponentNamespace = "component namespace switch"
+	ActionNone                            = "none"
+	ActionInstall                         = "install"
+	ActionUninstall                       = "uninstall"
+	ActionUpgrade                         = "upgrade"
+	ActionDowngrade                       = "downgrade"
+	ActionSwitchDoguNamespace             = "dogu namespace switch"
+	ActionSwitchComponentNamespace        = "component namespace switch"
+	ActionUpdateDoguProxyBodySize         = "update proxy body size"
+	ActionUpdateDoguProxyRewriteTarget    = "update proxy rewrite target"
+	ActionUpdateDoguProxyAdditionalConfig = "update proxy additional config"
+	ActionUpdateDoguResourceMinVolumeSize = "update resource minimum volume size"
 )