+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Dogu-specific:
 - [#11] apply new dogu states based on blueprint
 - [#20] Add exception for `nginx` dogu dependency validation. Map this dependency to `nginx-ingress` and `nginx-static`.
 - [#30] Implement dogu namespace switch.
+- [#55] Add dogu specific config for the volume size and the reverse proxy.
 
 Component-specific:
 - [#19] Create component differences between effective blueprint and cluster state